出品者でない場合トップページに戻る記述の追加

+++ app/controllers/items_controller.rb
@@ -18,6 +18,10 @@ class ItemsController < ApplicationController
   end
 
   def destroy
+    if user_signed_in? && current_user.id != @item.user_id
+      redirect_to :index
+    end
+    
     if @item.destroy
     redirect_to root_path
     end
